Implicit global admin is now explicitly endowed superadmin role (#148)
Previously, we consider an admin as global admin if the admin's domain is None, so an admin with domain name automatically recognized as domain admin.

After that, we introduced domain_name and group_id as required fields to kernels table to list kernels per domain or per group basis. This leads to the broken code execution for superadmin since kernel's domain_name field cannot be set.

To avoid this bug, and to explicitly recognize global admin, we added "superadmin" role to UserRole.

* Migration scripts append "superadmin" role to DB, and give default superadmin (admin@lablup.com) a default domain name and superadmin role.
* Global admin checking code is changed to inspect user's role instead of the emptiness of domain_name.
* For request handlers, request['is_superadmin'] is introduced in auth middleware.
* group_id field can be delivered by user when creating compute kernel.