Restructure RPC context manager slightly.
Honestly, I think it is a little bit complicated that I first open and close one Peer. Then, I have to open another Peer. Code looks dirty because of it. However, so far, it's the work around for letting establish Peer-to-Peer communication for each Agent-Manager pair with unique stream key.