feat: get current utilization considers every containers in a cluster session
If the target of the check_session is a cluster session, the current
utilization calculation considers sub-containers, by averaging them.

For example, the current utilization will be 30% for a cluster session with
the following utilizations for each container:
* main1: cpu-util: 50%
* sub1:  cpu-util: 10%

Note that I assumed that the allocated amount of resources are the same
across all the containers.